+++ library/src/methods/strict/strictAsync.ts
@@ -1,5 +1,5 @@
 import type { ObjectSchemaAsync } from '../../schemas/object/index.ts';
-import { getIssues } from '../../utils/index.ts';
+import { getSchemaIssues } from '../../utils/index.ts';
 
 /**
  * Creates a strict async object schema that throws an error if an input
@@ -29,7 +29,13 @@ export function strictAsync<TSchema extends ObjectSchemaAsync<any>>(
       const result = await schema._parse(input, info);
       return !result.issues &&
         Object.keys(input as object).some((key) => !(key in schema.object))
-        ? getIssues(info, 'object', 'strict', error || 'Invalid keys', input)
+        ? getSchemaIssues(
+            info,
+            'object',
+            'strict',
+            error || 'Invalid keys',
+            input
+          )
         : result;
     },
   };