+++ library/src/schemas/array/array.ts
@@ -1,5 +1,9 @@
 import type { BaseSchema, Input, Issues, Output, Pipe } from '../../types.ts';
-import { executePipe, getDefaultArgs, getIssues } from '../../utils/index.ts';
+import {
+  executePipe,
+  getDefaultArgs,
+  getSchemaIssues,
+} from '../../utils/index.ts';
 import type { ArrayPathItem } from './types.ts';
 
 /**
@@ -77,7 +81,13 @@ export function array<TArrayItem extends BaseSchema>(
     _parse(input, info) {
       // Check type of input
       if (!Array.isArray(input)) {
-        return getIssues(info, 'type', 'array', error || 'Invalid type', input);
+        return getSchemaIssues(
+          info,
+          'type',
+          'array',
+          error || 'Invalid type',
+          input
+        );
       }
 
       // Create issues and output