+++ library/src/schemas/array/arrayAsync.ts
@@ -9,7 +9,7 @@ import type {
 import {
   executePipeAsync,
   getDefaultArgs,
-  getIssues,
+  getSchemaIssues,
 } from '../../utils/index.ts';
 import type { ArrayPathItem } from './types.ts';
 
@@ -88,7 +88,13 @@ export function arrayAsync<TArrayItem extends BaseSchema | BaseSchemaAsync>(
     async _parse(input, info) {
       // Check type of input
       if (!Array.isArray(input)) {
-        return getIssues(info, 'type', 'array', error || 'Invalid type', input);
+        return getSchemaIssues(
+          info,
+          'type',
+          'array',
+          error || 'Invalid type',
+          input
+        );
       }
 
       // Create issues and output