+++ library/src/schemas/nonNullable/nonNullableAsync.ts
@@ -4,7 +4,7 @@ import type {
   Input,
   Output,
 } from '../../types.ts';
-import { getIssues } from '../../utils/index.ts';
+import { getSchemaIssues } from '../../utils/index.ts';
 import type { NonNullable } from './nonNullable.ts';
 
 /**
@@ -57,7 +57,7 @@ export function nonNullableAsync<TWrapped extends BaseSchema | BaseSchemaAsync>(
     async _parse(input, info) {
       // Allow `null` values not to pass
       if (input === null) {
-        return getIssues(
+        return getSchemaIssues(
           info,
           'type',
           'non_nullable',