+++ library/src/schemas/nonNullish/nonNullish.ts
@@ -1,5 +1,5 @@
 import type { BaseSchema, Input, Output } from '../../types.ts';
-import { getIssues } from '../../utils/index.ts';
+import { getSchemaIssues } from '../../utils/index.ts';
 
 /**
  * Non nullish type.
@@ -56,7 +56,7 @@ export function nonNullish<TWrapped extends BaseSchema>(
     _parse(input, info) {
       // Allow `null` and `undefined` values not to pass
       if (input === null || input === undefined) {
-        return getIssues(
+        return getSchemaIssues(
           info,
           'type',
           'non_nullish',